// src/features/mutual-funds-analysis/lib/types.ts
export type MutualFund = {
	schemeCode: number;
	schemeName: string;
	earliestDate: string;
	lastDate: string;
	navs: Map<string, number>;
};

export type MutualFundRollingReturn = {
	noData: boolean;
	avgReturn: number;
	minReturn: number;
	maxReturn: number;
};

export type MutualFundReturn = {
	schemeCode: number;
	date: string;
	return: number;
};

export enum PresetTimeDurations {
	OneMonth = '1Month',
	TwoMonths = '2Months',
	ThreeMonths = '3Months',
	SixMonths = '6Months',
	OneYear = '1Year',
	TwoYears = '2Years',
	ThreeYears = '3Years',
	FiveYears = '5Years',
	TenYears = '10Years',
	FifteenYears = '15Years',
	TwentyYears = '20Years',
}